renamed "components" folder to "_components"
